requirements-dev.txt
jsonschema>=4.19
PyYAML>=6.0
python-dateutil>=2.9
tabulate>=0.9

